Split script into modules for expandablity
